Make Extended RPC interface plugin based only
