+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ All notable changes to this project are documented in this file.
 
 [0.8.3-dev] in progress
 -----------------------
--
+- Make the ExtendedRPC interface plugin based only.
 
 
 [0.8.2] 2018-10-31
