+++ neo/api/JSONRPC/ExtendedJsonRpcApi.py
@@ -1,8 +1,38 @@
-from neo.Core.Blockchain import Blockchain
-from neo.api.JSONRPC.JsonRpcApi import JsonRpcApi, JsonRpcError
-from neo.Implementations.Wallets.peewee.UserWallet import UserWallet
-from neocore.UInt256 import UInt256
-import datetime
+import os
+import inspect
+import pkgutil
+from importlib import import_module
+from neo.api.JSONRPC.JsonRpcApi import JsonRpcApi
+from neo.api.JSONRPC.ExtendedRpcCommand import ExtendedRpcCommand
+import neo.api.JSONRPC.extended_commands as rpc_plugins
+
+plugins = []
+
+
+def iter_namespace(ns_pkg):
+    return pkgutil.iter_modules(ns_pkg.__path__, ns_pkg.__name__ + ".")
+
+
+# search in the plugins dir
+for (importer, _name, ispkg) in iter_namespace(rpc_plugins):
+    # only for plugins in packages
+    if ispkg:
+        # load the package
+        m = importer.find_module(_name).load_module(_name)
+        # and iterate over its contents
+        for (_, mod_name, _) in iter_namespace(m):
+            try:
+                imported_module = import_module(mod_name)
+
+                for i in dir(imported_module):
+                    attribute = getattr(imported_module, i)
+
+                    # to find a class that extends ExtendedRpcCommand
+                    if inspect.isclass(attribute) and issubclass(attribute, ExtendedRpcCommand):
+                        plugins.append(attribute)
+
+            except ImportError:
+                pass
 
 
 class ExtendedJsonRpcApi(JsonRpcApi):
@@ -11,51 +41,15 @@ class ExtendedJsonRpcApi(JsonRpcApi):
     """
 
     def __init__(self, port, wallet=None):
-        self.start_height = Blockchain.Default().Height
-        self.start_dt = datetime.datetime.utcnow()
         super(ExtendedJsonRpcApi, self).__init__(port, wallet)
+        self.commands = {}
+        for plugin in plugins:
+            for command in plugin.commands():
+                self.commands.update({command: plugin})
 
     def json_rpc_method_handler(self, method, params):
-
-        if method == "getnodestate":
-            height = Blockchain.Default().Height
-            headers = Blockchain.Default().HeaderHeight
-            diff = height - self.start_height
-            now = datetime.datetime.utcnow()
-            difftime = now - self.start_dt
-            mins = difftime / datetime.timedelta(minutes=1)
-            secs = mins * 60
-            bpm = 0
-            tps = 0
-
-            if diff > 0 and mins > 0:
-                bpm = diff / mins
-                tps = Blockchain.Default().TXProcessed / secs
-
-            return {
-                'Progress': [height, "/", headers],
-                'Block-cache length': Blockchain.Default().BlockCacheCount,
-                'Blocks since program start': diff,
-                'Time elapsed (minutes)': mins,
-                'Blocks per min': bpm,
-                'TPS': tps
-            }
-
-        elif method == "gettxhistory":
-            if self.wallet:
-                res = []
-                for tx in self.wallet.GetTransactions():
-                    json = tx.ToJson()
-                    tx_id = UInt256.ParseString(json['txid'])
-                    txx, height = Blockchain.Default().GetTransaction(tx_id)
-                    header = Blockchain.Default().GetHeaderByHeight(height)
-                    block_index = header.Index
-                    json['block_index'] = block_index
-                    block_timestamp = header.Timestamp
-                    json['blocktime'] = block_timestamp
-                    res.append(json)
-                return res
-            else:
-                raise JsonRpcError(-400, "Access denied.")
-
-        return super(ExtendedJsonRpcApi, self).json_rpc_method_handler(method, params)  
+        plugin = self.commands.get(method)
+        if plugin:
+            return plugin.execute(self, method, params)
+        else:
+            return super(ExtendedJsonRpcApi, self).json_rpc_method_handler(method, params)
