+++ neo/api/JSONRPC/ExtendedRpcCommand.py
@@ -1,0 +1,13 @@
+from abc import ABC, abstractmethod
+
+
+class ExtendedRpcCommand(ABC):
+    @classmethod
+    @abstractmethod
+    def commands(cls):
+        return []
+
+    @classmethod
+    @abstractmethod
+    def execute(cls, json_rpc_api, method):
+        return "result"
